requirements.txt
streamlit==1.42.0
google-genai==0.3.0
streamlit-drawable-canvas==0.9.3
requests==2.32.3
pydantic==2.10.4
python-dotenv==1.0.1
toml==0.10.2
